--- src/laser_setup/__main__.py
@@ -1,13 +1,0 @@
-"""
-Trying some stuff out.
-TODO:
-    - Fix the folder structure
-    - Fix the imports
-    - Fix tests
-    - Decide on a logging strategy
-    - Decide on a config strategy
-    - Decide on a data storage strategy
-"""
-from src.lib.utils import config
-
-breakpoint()
