--- src/lib/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ['fits', 'k2450', 'utils']
