+++ tests/ramp_test.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 
-from src.lib.utils import gate_sweep_ramp
+from lib.utils import gate_sweep_ramp
 
 def test_ramp():
     ramp = gate_sweep_ramp(-35, 35, 0.5)
